Push versioning to master

+++ catapult-sdk/src/modelBinary/verifiableEntityCodec.js
@@ -33,7 +33,7 @@ const verifiableEntityCodec = {
 		const entity = {};
 		entity.signature = parser.buffer(constants.sizes.signature);
 		entity.signer = parser.buffer(constants.sizes.signer);
-		entity.version = parser.uint16();
+		entity.version = parser.uint32();
 		entity.type = parser.uint16();
 		return entity;
 	},
@@ -46,7 +46,7 @@ const verifiableEntityCodec = {
 	serialize: (entity, serializer) => {
 		serializer.writeBuffer(entity.signature);
 		serializer.writeBuffer(entity.signer);
-		serializer.writeUint16(entity.version);
+		serializer.writeUint32(entity.version);
 		serializer.writeUint16(entity.type);
 	}
 };
